Link to/from Frog pages added, new icon added
We need to add this new icon to the icon directory on the Frog web
server, and the JS needs to be updated on both the Internal and Frog
side now. Updated Frog config for adding a "back" link in top right to
the previous page the user was on in the iframe.