Switched to googlechartapi
Harvey Simon really wanted to be able to include the graph in his weekly
report, but the default chart has no scale bars, so he had to recreate
the chart manually in Excel.  Switched to using the googlechartsapi
plugin, but needed to customize it somewhat because we weren't showing
page visitors, only hits.  Also copied plugins folder to annual
subfolder for completeness.